create a factory base class and use it in the same format across use cases
